// src/app/model/sort-option.ts
/**
 * 
 * Represents the options used in dropdown selects when sorting.
 * @interface SortOption
 * @property {string} value - The value of the sort option.
 * @property {string} viewValue - The displayed value of the sort option.
 */
export interface SortOption {
    value: string;
    viewValue: string;
}